requirements.txt
# Meta Ads Optimizer - Dependencies

# Meta/Facebook API
facebook-business>=22.0.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# Anthropic API (Agent Brain)
anthropic>=0.39.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
# SQLite is built into Python, no extra package needed

# Utilities
requests>=2.28.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
ruff>=0.1.0
